backend/requirements.txt
--extra-index-url https://download.pytorch.org/whl/cpu

deep_translator==1.11.4
fastapi==0.115.12
groq==0.22.0
gTTS==2.5.4
lxml==5.3.0
lxml_html_clean
nltk==3.9.1
numpy<2
pandas==2.2.3
python-dotenv==1.1.0
requests==2.32.3
seaborn==0.13.2
trafilatura==2.0.0
transformers==4.50.3
uvicorn==0.34.0

# PyTorch (CPU-only versions)
torch==2.2.1
torchvision==0.17.1
torchaudio==2.2.1
